Separate file event collector
from ContinuousBuildTriggerHandler
