+++ subprojects/launcher/src/main/java/org/gradle/tooling/internal/provider/ContinuousBuildTriggerHandler.java
@@ -16,54 +16,43 @@
 
 package org.gradle.tooling.internal.provider;
 
-import org.gradle.execution.plan.BuildInputHierarchy;
 import org.gradle.initialization.BuildCancellationToken;
 import org.gradle.initialization.ContinuousExecutionGate;
 import org.gradle.internal.UncheckedException;
 import org.gradle.internal.filewatch.PendingChangesListener;
-import org.gradle.internal.logging.text.StyledTextOutput;
-import org.gradle.internal.os.OperatingSystem;
-import org.gradle.internal.watch.registry.FileWatcherRegistry;
-import org.gradle.internal.watch.vfs.FileChangeListener;
+import org.slf4j.Logger;
+import org.slf4j.LoggerFactory;
 
-import java.nio.file.Files;
-import java.nio.file.Path;
-import java.util.LinkedHashMap;
-import java.util.Map;
 import java.util.concurrent.CountDownLatch;
 import java.util.concurrent.TimeUnit;
 import java.util.concurrent.atomic.AtomicBoolean;
 
 import static org.gradle.internal.filewatch.DefaultFileSystemChangeWaiterFactory.QUIET_PERIOD_SYSPROP;
 
-public class ContinuousBuildTriggerHandler implements FileChangeListener {
-    private static final boolean IS_MAC_OSX = OperatingSystem.current().isMacOsX();
+public class ContinuousBuildTriggerHandler {
+    private static final Logger LOGGER = LoggerFactory.getLogger(ContinuousBuildTriggerHandler.class);
 
-    private final BuildInputHierarchy buildInputHierarchy;
     private final PendingChangesListener pendingChangesListener;
     private final BuildCancellationToken cancellationToken;
     private final ContinuousExecutionGate continuousExecutionGate;
     private final AtomicBoolean changeArrived = new AtomicBoolean(false);
     private final CountDownLatch changeOrCancellationArrived = new CountDownLatch(1);
     private final CountDownLatch cancellationArrived = new CountDownLatch(1);
-    private final FileEventCollector fileEventCollector = new FileEventCollector();
     private final long quietPeriod;
     private volatile long lastChangeAt = monotonicClockMillis();
 
     public ContinuousBuildTriggerHandler(
-        BuildInputHierarchy buildInputHierarchy,
         PendingChangesListener pendingChangesListener,
         BuildCancellationToken cancellationToken,
         ContinuousExecutionGate continuousExecutionGate
     ) {
-        this.buildInputHierarchy = buildInputHierarchy;
         this.pendingChangesListener = pendingChangesListener;
         this.cancellationToken = cancellationToken;
         this.continuousExecutionGate = continuousExecutionGate;
         this.quietPeriod = Long.getLong(QUIET_PERIOD_SYSPROP, 250L);
     }
 
-    void wait(Runnable notifier) {
+    public void wait(Runnable notifier) {
         Runnable cancellationHandler = () -> {
             changeOrCancellationArrived.countDown();
             cancellationArrived.countDown();
@@ -93,103 +82,18 @@ public class ContinuousBuildTriggerHandler implements FileChangeListener {
         }
     }
 
-    public void reportChanges(StyledTextOutput logger) {
-        fileEventCollector.reportChanges(logger);
-    }
-
-    @Override
-    public void handleChange(FileWatcherRegistry.Type type, Path path) {
-        String absolutePath = path.toString();
+    public void notifyFileChangeArrived() {
         lastChangeAt = monotonicClockMillis();
-        if (buildInputHierarchy.isInput(absolutePath)) {
-            // got a change, store it
-            fileEventCollector.onChange(type, path);
-            notifyChangeArrived();
-        }
-    }
-
-    @Override
-    public void stopWatchingAfterError() {
-        fileEventCollector.errorWhenWatching();
-        notifyChangeArrived();
-    }
-
-    private void notifyChangeArrived() {
         changeOrCancellationArrived.countDown();
         if (changeArrived.compareAndSet(false, true)) {
+            LOGGER.debug("notifying pending changes");
             pendingChangesListener.onPendingChanges();
+        } else {
+            LOGGER.debug("already notified");
         }
     }
 
     private static long monotonicClockMillis() {
         return TimeUnit.NANOSECONDS.toMillis(System.nanoTime());
-    }
-
-    private static class FileEventCollector {
-        private static final int SHOW_INDIVIDUAL_CHANGES_LIMIT = 3;
-        private final Map<Path, FileWatcherRegistry.Type> aggregatedEvents = new LinkedHashMap<>();
-        private int moreChangesCount;
-        private boolean errorWhenWatching;
-
-
-        public void onChange(FileWatcherRegistry.Type type, Path path) {
-            FileWatcherRegistry.Type existingEvent = aggregatedEvents.get(path);
-            if (existingEvent == type ||
-                (existingEvent == FileWatcherRegistry.Type.CREATED && type == FileWatcherRegistry.Type.MODIFIED)) {
-                return;
-            }
-
-            if (existingEvent != null || aggregatedEvents.size() < SHOW_INDIVIDUAL_CHANGES_LIMIT) {
-                aggregatedEvents.put(path, type);
-            } else if (shouldIncreaseChangesCount(type, path)) {
-                moreChangesCount++;
-            }
-        }
-
-        private boolean shouldIncreaseChangesCount(FileWatcherRegistry.Type type, Path path) {
-            // Count every event on macOS, since there is only one event for file creation.
-            return IS_MAC_OSX ||
-                // On other operating systems count only non-CREATE events, since creation also causes a modification event, unless the event is for a directory.
-                type != FileWatcherRegistry.Type.CREATED || Files.isDirectory(path);
-        }
-
-
-        public void errorWhenWatching() {
-            errorWhenWatching = true;
-        }
-
-        public void reportChanges(StyledTextOutput logger) {
-            for (Map.Entry<Path, FileWatcherRegistry.Type> entry : aggregatedEvents.entrySet()) {
-                FileWatcherRegistry.Type type = entry.getValue();
-                Path path = entry.getKey();
-                showIndividualChange(logger, path, type);
-            }
-            if (moreChangesCount > 0) {
-                logOutput(logger, "and some more changes");
-            }
-            if (errorWhenWatching) {
-                logOutput(logger, "Error when watching files - triggering a rebuild");
-            }
-        }
-
-        private void showIndividualChange(StyledTextOutput logger, Path path, FileWatcherRegistry.Type changeType) {
-            String changeDescription;
-            switch (changeType) {
-                case CREATED:
-                    changeDescription = "new " + (Files.isDirectory(path) ? "directory" : "file");
-                    break;
-                case REMOVED:
-                    changeDescription = "deleted";
-                    break;
-                case MODIFIED:
-                default:
-                    changeDescription = "modified";
-            }
-            logOutput(logger, "%s: %s", changeDescription, path.toString());
-        }
-
-        private void logOutput(StyledTextOutput logger, String message, Object... objects) {
-            logger.formatln(message, objects);
-        }
     }
 }
